+++ src/evm/chain/exec.rs
@@ -1,7 +1,6 @@
 //! Execution output type.
 
-use crate::evm::chain::Transaction;
-use crate::evm::cheatcode::BrokenInvariant;
+use crate::evm::chain::BrokenInvariant;
 use crate::evm::{ExecutionCoverage, result, trace};
 
 /// Result of executing a sequence of transactions.
@@ -10,6 +9,5 @@ pub struct ExecOutput {
     pub results: Vec<result::TransactionResult>,
     pub trace: Option<trace::Trace>,
     pub coverage: Option<ExecutionCoverage>,
-    pub panic_transactions: Vec<Transaction>,
     pub broken_invariants: Vec<Vec<BrokenInvariant>>,
 }