+++ src/evm/cheatcode/mod.rs
@@ -6,7 +6,7 @@
 pub use address::VM_ADDRESS;
 pub use config::CheatcodeConfig;
 pub use inspector::Inspector;
-pub use state::{BrokenInvariant, ExecutionState};
+pub use state::ExecutionState;
 
 mod address;
 mod config;